backend/requirements.txt
fastapi
python-socketio[asgi]
uvicorn
httpx